requirements.txt
pyrogram==2.0.106
telegraph
Tgcrypto
aiohttp
pymongo[srv]==3.12.3
motor==2.5.1
marshmallow==3.14.1
umongo==3.0.1
validators
requests
beautifulsoup4
pytz
python-dotenv
Flask
gunicorn
Jinja2
Werkzeug
itsdangerous
pyrofork

#---ytdl----# 
ffmpeg
youtube_search
youtube-search-python==1.4.6
yt-dlp
youtube-dl
requests 
wget
